Consolidate realm into one folder
