+++ app/realm/actions/general.js
@@ -3,8 +3,8 @@
 
 import {Client4} from 'mattermost-redux/client';
 
-import {GeneralTypes} from 'app/action_types';
-import {GENERAL_SCHEMA_ID} from 'app/models/general';
+import {GeneralTypes} from 'app/realm/action_types';
+import {GENERAL_SCHEMA_ID} from 'app/realm/models/general';
 import PushNotifications from 'app/push_notifications';
 import ephemeralStore from 'app/store/ephemeral_store';
 import {t} from 'app/utils/i18n';
