+++ app/realm/models/channel.js
