+++ app/realm/models/emoji.js
