+++ app/realm/utils/general.js
