+++ app/screens/login/index.js
@@ -3,8 +3,8 @@
 
 import {realmConnect} from 'realm-react-redux';
 
-import {scheduleExpiredNotification, sendPasswordResetEmail} from 'app/actions/realm/general';
-import {login} from 'app/actions/realm/user';
+import {scheduleExpiredNotification, sendPasswordResetEmail} from 'app/realm/actions/general';
+import {login} from 'app/realm/actions/user';
 import options from 'app/store/realm_context_options';
 import {reduxStore} from 'app/store';
